Split EventDefinitionsPage into smaller components
