+++ graylog2-web-interface/src/pages/EventDefinitionsPage.jsx
@@ -1,88 +1,15 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import { LinkContainer } from 'react-router-bootstrap';
-import { Button, ButtonToolbar, Col, DropdownButton, MenuItem, Row } from 'react-bootstrap';
-import lodash from 'lodash';
+import { Button, ButtonToolbar, Col, Row } from 'react-bootstrap';
 
-import { DocumentTitle, EmptyEntity, EntityList, EntityListItem, PageHeader } from 'components/common';
+import { DocumentTitle, PageHeader } from 'components/common';
 import DocumentationLink from 'components/support/DocumentationLink';
+import EventDefinitionsContainer from 'components/event-definitions/event-definitions/EventDefinitionsContainer';
 
-import connect from 'stores/connect';
 import Routes from 'routing/Routes';
 import DocsHelper from 'util/DocsHelper';
-import CombinedProvider from 'injection/CombinedProvider';
-
-const { EventDefinitionsStore, EventDefinitionsActions } = CombinedProvider.get('EventDefinitions');
 
 class EventDefinitionsPage extends React.Component {
-  static propTypes = {
-    eventDefinitions: PropTypes.object.isRequired,
-  };
-
-  componentDidMount() {
-    EventDefinitionsActions.list();
-  }
-
-  handleDelete = (definition) => {
-    return () => {
-      if (window.confirm(`Are you sure you want to delete "${definition.title}"?`)) {
-        EventDefinitionsActions.delete(definition);
-      }
-    };
-  };
-
-  renderEmptyContent = () => {
-    return (
-      <Row>
-        <Col md={4} mdOffset={4}>
-          <EmptyEntity>
-            <p>
-              Create Event Definitions that are able to search, aggregate or correlate Messages and other
-              Events, allowing you to record significant Events in Graylog and alert on them.
-            </p>
-            <LinkContainer to={Routes.NEXT_ALERTS.DEFINITIONS.CREATE}>
-              <Button bsStyle="success">Get Started!</Button>
-            </LinkContainer>
-          </EmptyEntity>
-        </Col>
-      </Row>
-    );
-  };
-
-  renderContent = () => {
-    const { eventDefinitions } = this.props;
-    const definitions = eventDefinitions.list;
-
-    if (eventDefinitions.list.length < 1) {
-      return this.renderEmptyContent();
-    }
-
-    const items = definitions.map((definition) => {
-      const actions = [
-        <LinkContainer key={`edit-button-${definition.id}`}
-                       to={Routes.NEXT_ALERTS.DEFINITIONS.edit(definition.id)}>
-          <Button bsStyle="info">Edit</Button>
-        </LinkContainer>,
-        <DropdownButton key={`actions-${definition.id}`} id="more-dropdown" title="More" pullRight>
-          <MenuItem onClick={this.handleDelete(definition)}>Delete</MenuItem>
-        </DropdownButton>,
-      ];
-
-      // TODO: Show something more useful ;)
-      const titleSuffix = lodash.get(definition, 'config.type', 'Not available')
-        .replace(/-v\d+/, '');
-      return (
-        <EntityListItem key={`event-definition-${definition.id}`}
-                        title={definition.title}
-                        titleSuffix={`${titleSuffix}`}
-                        description={definition.description}
-                        actions={actions} />
-      );
-    });
-
-    return (<EntityList items={items} />);
-  };
-
   render() {
     return (
       <DocumentTitle title="Event Definitions">
@@ -112,7 +39,7 @@ class EventDefinitionsPage extends React.Component {
             <Col md={12}>
               <Row>
                 <Col md={12}>
-                  {this.renderContent()}
+                  <EventDefinitionsContainer />
                 </Col>
               </Row>
             </Col>
@@ -123,4 +50,4 @@ class EventDefinitionsPage extends React.Component {
   }
 }
 
-export default connect(EventDefinitionsPage, { eventDefinitions: EventDefinitionsStore });
+export default EventDefinitionsPage;
